Started to incorporate the GUI
Still unfinished, this is my first time using PyQt and working with GUIs so it doesn't work properly yet. I am using the Model-Controller-View pattern to separate my code. So far I have made the main window but not yet the window to display the results. The submit button is also not working yet but hopefully soon I will figure out the issue.